Introduce `RoomOpener`
