+++ apps/meteor/client/views/omnichannel/directory/calls/Call.tsx
@@ -1,23 +1,16 @@
 import type { IRoom } from '@rocket.chat/core-typings';
 import { Box } from '@rocket.chat/fuselage';
 import type { ReactElement } from 'react';
-import React, { Suspense } from 'react';
+import React from 'react';
 
-import { useOpenRoom } from '../../../../hooks/useOpenRoom';
-import { Room, RoomNotFound, RoomProvider, RoomSkeleton } from '../../../room';
+import RoomOpener from '../../../room/RoomOpener';
 
 type CallProps = { rid: IRoom['_id'] };
 
 const Call = ({ rid }: CallProps): ReactElement => {
-	const { data, isSuccess, isError, isLoading } = useOpenRoom({ type: 'v', ref: rid });
-
 	return (
 		<Box position='absolute' backgroundColor='surface' width='full' height='full'>
-			<Suspense fallback={<RoomSkeleton />}>
-				{isLoading && <RoomSkeleton />}
-				{isSuccess && ('rid' in data ? <RoomProvider rid={data.rid}>{<Room />}</RoomProvider> : <RoomSkeleton />)}
-				{isError && <RoomNotFound />}
-			</Suspense>
+			<RoomOpener type='v' reference={rid} />
 		</Box>
 	);
 };
